+++ modules/core/dependency/python-ihm/test/test_examples.py
@@ -24,11 +24,11 @@ class Tests(unittest.TestCase):
     @unittest.skipIf('APPVEYOR' in os.environ,
                      "AppVeyor environments have old SSL certs")
     @unittest.skipIf('GITHUB_ACTIONS' in os.environ,
-                     "Example is slow and fails when PDB-Dev is down")
+                     "Example is slow and fails when PDB-IHM is down")
     def test_validator_example(self):
         """Test validator example"""
         subprocess.check_call([sys.executable,
-                               get_example_path("validate_pdb_dev.py")])
+                               get_example_path("validate_pdb_ihm.py")])
 
     def test_simple_docking_example(self):
         """Test simple-docking example"""